requirements.txt
annotated-types==0.6.0
anyio==3.6.2
bcrypt==4.0.1
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
dnspython==2.3.0
docopt==0.6.2
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.2.0
Faker==20.1.0
fastapi==0.95.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
passlib==1.7.4
pipreqs==0.4.13
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
pydantic_core==2.14.5
PyJWT==2.6.0
pymongo==4.3.3
pyOpenSSL==22.1.0
python-dateutil==2.8.2
python-decouple==3.8
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.6.0
urllib3==2.1.0
uvicorn==0.22.0
yarg==0.1.9
